+++ lib/src/utils.dart
@@ -2,6 +2,7 @@ library utils;
 
 import 'package:date/date.dart';
 import 'package:timeseries/timeseries.dart';
+import 'package:timezone/timezone.dart';
 
 /// Merge a map of timeseries.  Only values that exist are kept.
 TimeSeries<Map<K, T>> mergeAll<K, T>(Map<K, TimeSeries<T>> xs) {
@@ -19,6 +20,59 @@ TimeSeries<Map<K, T>> mergeAll<K, T>(Map<K, TimeSeries<T>> xs) {
         if (y != null) x[keys[k]] = y;
         return x;
       }, joinType: JoinType.Outer);
+    }
+  }
+  return out;
+}
+
+/// Groups an hourly/sub-hourly time series by a specified hour range.
+/// The keys of the returned map are intervals corresponding to the
+/// specified hour range.
+///
+/// The [hourRange] is a tuple of two integers representing the start and end
+/// hours.  If the start hour is after the end hour, the interval spills into
+/// the following day.  The [hourRange] is inclusive of both the start and end 
+/// hours.  If the start and end hours are the same, the interval spans a 
+/// full day.  
+/// 
+/// Values for the start/end hours should be between 0 and 23.
+///
+/// Note that you'll likely have incomplete groups at the start and end of the
+/// time series.
+///
+Map<Interval, TimeSeries<K>> groupByHourRange<K>(
+    TimeSeries<K> ts, (int, int) hourRange) {
+  assert(hourRange.$1 >= 0 && hourRange.$1 <= 23);
+  assert(hourRange.$2 >= 0 && hourRange.$2 <= 23);
+
+  var out = <Interval, TimeSeries<K>>{};
+  final startHour = hourRange.$1;
+  final endHour = hourRange.$2;
+  for (var e in ts) {
+    var start = TZDateTime(e.interval.start.location, e.interval.start.year,
+        e.interval.start.month, e.interval.start.day, startHour);
+    var end = TZDateTime(e.interval.start.location, e.interval.start.year,
+        e.interval.start.month, e.interval.start.day, endHour);
+    if (startHour > endHour) {
+      if (e.interval.start.hour >= startHour) {
+        end = TZDateTime(e.interval.start.location, e.interval.start.year,
+                e.interval.start.month, e.interval.start.day, endHour)
+            .add(const Duration(days: 1));
+      } else {
+        start = TZDateTime(e.interval.start.location, e.interval.start.year,
+                e.interval.start.month, e.interval.start.day, startHour)
+            .subtract(const Duration(days: 1));
+      }
+    } else if (startHour == endHour) {
+      if (e.interval.start.isBefore(start)) {
+        start = start.subtract(const Duration(days: 1));
+      } else {
+        end = start.add(const Duration(days: 1));
+      }
+    }
+    final key = Interval(start, end);
+    if (key.containsInterval(e.interval)) {
+      out.putIfAbsent(key, () => TimeSeries<K>()).add(e);
     }
   }
   return out;