+++ lib/timeseries.dart
@@ -4,7 +4,6 @@ export 'src/check_missing.dart';
 export 'src/common_aggregations.dart';
 export 'src/group_term.dart';
 export 'src/interval_tuple.dart';
-export 'src/time_tuple.dart';
 export 'src/timeseries_base.dart';
 export 'src/timeseries_ext.dart';
 export 'src/numeric_timeseries_ext.dart';